lift #1 day 11: sunset OLD exporters (full move + delete)
Per features/03_export/basevla-spine_plan.md Day 11. The legacy
exporter modules contained load-bearing builders that the new spine
paths depend on. Day 11 moves those builders into the spine exporter
modules and deletes the legacy files outright.

Deleted (gone from main):

- src/reflex/exporters/smolvla_exporter.py (306 LOC)
- src/reflex/exporters/gr00t_exporter.py    (891 LOC)

Renamed:

- src/reflex/exporters/pi0_exporter.py → src/reflex/exporters/pi0.py
  (history preserved via git mv; content unchanged at the file level)

Spine modules absorbed the load-bearing pieces:

- src/reflex/exporters/smolvla.py gained:
  * `build_expert_stack` (from legacy smolvla_exporter)
  * Re-exports `ExpertGQALayer` + `ExpertStack` from
    `reflex.models.heads.expert_stack` (preserves external import surface)

- src/reflex/exporters/gr00t.py gained:
  * `GR00T_BLOCK_PREFIX` + `GR00T_META_KEYS` constants
  * `_sinusoidal_timestep` helper
  * `GR00TDiTBlock` + `GR00TExpertStack` classes
  * `GR00TActionEncoder` + `GR00TStateEncoder` + `GR00TActionDecoder` +
    `GR00TFullStack` classes
  * `build_gr00t_expert_stack` + `build_gr00t_full_stack` builders

Updated import sites (15+ files across src/ + tests/ + scripts/):

- src/reflex/_pytorch_backend.py: 3 imports
- src/reflex/cli.py: pi0 + smolvla + gr00t paths now reference renamed
  / consolidated modules
- src/reflex/exporters/monolithic.py: gr00t import
- src/reflex/exporters/pi0_prefix.py: 3 pi0 imports
- src/reflex/models/heads/dit_head.py: gr00t import
- src/reflex/models/heads/flow_matching_head.py: pi0/pi05/smolvla imports
- src/reflex/models/vlas/pi0.py + smolvla.py: spine builders
- tests/test_checkpoint_detection.py: 4 pi0 imports
- tests/test_flow_matching_head.py: smolvla module reference
- tests/test_smolvla_spine.py: 2 smolvla imports
- scripts/modal_bench_* + modal_test_* + modal_gr00t_* + modal_verify_*:
  bulk-updated via sed

New test:

- tests/test_day10_cli_vla_type.py::test_legacy_exporter_modules_deleted
  pins that pi0_exporter / smolvla_exporter / gr00t_exporter are GONE
  (importlib raises ModuleNotFoundError). Catches accidental
  re-introduction during future refactors.

Acceptance gates (plan Day 11):

- src/reflex/runtime/ has zero references to legacy modules ✓
  (verified earlier; runtime was already independent of exporter paths)
- All Day 4-10 spine tests + flow_matching_head + registry_completeness
  + checkpoint_detection green: 96 pass / 0 fail
- LIBERO regression + 9× speedup smoke deferred to a follow-up Modal
  fire (per the existing Day 9 acceptance deferral pattern)
