+++ scripts/modal_bench_onnx_vs_torch.py
@@ -55,28 +55,28 @@ def bench():
         {
             "tag": "smolvla",
             "hf_id": "lerobot/smolvla_base",
-            "builder_mod": "reflex.exporters.smolvla_exporter",
+            "builder_mod": "reflex.exporters.smolvla",
             "builder_fn": "build_expert_stack",
             "builder_kwargs": {"head_dim": 64},
         },
         {
             "tag": "pi0",
             "hf_id": "lerobot/pi0_base",
-            "builder_mod": "reflex.exporters.pi0_exporter",
+            "builder_mod": "reflex.exporters.pi0",
             "builder_fn": "build_pi0_expert_stack",
             "builder_kwargs": {"head_dim": 128},
         },
         {
             "tag": "pi05",
             "hf_id": "lerobot/pi05_base",
-            "builder_mod": "reflex.exporters.pi0_exporter",
+            "builder_mod": "reflex.exporters.pi0",
             "builder_fn": "build_pi05_expert_stack",
             "builder_kwargs": {"head_dim": 128},
         },
         {
             "tag": "gr00t",
             "hf_id": "nvidia/GR00T-N1.6-3B",
-            "builder_mod": "reflex.exporters.gr00t_exporter",
+            "builder_mod": "reflex.exporters.gr00t",
             "builder_fn": "build_gr00t_full_stack",
             "builder_kwargs": {"embodiment_id": 0},
         },
