+++ scripts/modal_gr00t_vlm_parity.py
@@ -80,7 +80,7 @@ def run_parity(model_id: str = "nvidia/GR00T-N1.6-3B"):
     import torch
 
     from reflex.checkpoint import load_checkpoint
-    from reflex.exporters.gr00t_exporter import build_gr00t_full_stack
+    from reflex.exporters.gr00t import build_gr00t_full_stack
 
     print(f"[parity] Loading {model_id}...")
     t0 = time.time()
@@ -118,7 +118,7 @@ def run_parity(model_id: str = "nvidia/GR00T-N1.6-3B"):
     # ours.dit, ours.action_decoder) but wires them directly, without
     # going through GR00TFullStack.forward. Any delta = wrapper bug.
     import torch.nn.functional as F
-    from reflex.exporters.gr00t_exporter import _sinusoidal_timestep
+    from reflex.exporters.gr00t import _sinusoidal_timestep
 
     print(f"\n[parity] Path B (hand-rolled reference, same primitives)")
     with torch.no_grad():
