+++ scripts/modal_test_pi0.py
@@ -71,7 +71,7 @@ def run_pi0_export():
     print("\n=== Step 2: Build pi0 expert stack ===", flush=True)
     start = time.time()
     try:
-        from reflex.exporters.pi0_exporter import build_pi0_expert_stack
+        from reflex.exporters.pi0 import build_pi0_expert_stack
         expert_stack, meta = build_pi0_expert_stack(state_dict, head_dim=128)
         elapsed = time.time() - start
         log("build_expert", "pass",
